Fix rendering of initial Toggle setting state.
Setting values are typically strings, but the ToggleSettingSpecifier
trueValue is typically a Boolean. Thus the eq comparison fails when
rendering the toggle HTML because `"true" eq true` returns false.

Fixed by adding a helper method isToggleSettingValueTrue() to coerce the
setting value into a Boolean for the comparison, falling back to
coercing the toggle trueValue to a string and performing a string
comparison if the trueValue is not a Boolean.