Fix parsing of Unit ID and Block Type columns from database as ints.
The Unit ID and Block Type register values are stored in SMALLINT
database columns, but were being parsed as a Byte type when reading.
This worked for values < 127 but for values larger an exception was
thrown because bytes are signed in Java. The fix is to read the columns
as integers, which they were being cast to anyway.